+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:example/reactive_home_example.dart';
 import 'package:flutter/material.dart';
 import 'package:xlerate/size_config.dart';
 import 'package:xlerate/xlerate.dart';
@@ -24,7 +25,7 @@ class MyApp extends StatelessWidget {
         ),
         useMaterial3: true,
       ),
-      home: const HomePage(),
+      home: const ReactiveHomeExample(),
     );
   }
 }