+++ example/lib/reactive_home_example.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:xlerate/reactive_widgets.dart';
+import 'package:xlerate/xlerate_web.dart';
 
 class ReactiveHomeExample extends StatelessWidget {
   const ReactiveHomeExample({Key? key}) : super(key: key);
@@ -45,15 +45,15 @@ class MyWidget2 extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
+    return Scaffold(
       body: Center(
-        child: Text(
+        child: const Text(
           'This is Widget2',
           style: TextStyle(
             fontSize: 16,
             fontWeight: FontWeight.bold,
           ),
-        ),
+        ).txt.create(),
       ),
     );
   }