+++ lib/reactive_widgets.dart
@@ -1,7 +1,0 @@
-///Importing 'reactive_widgets' will only give you access to `reactive_row_col` and `reactive_home`.
-library reactive_widgets;
-
-//From Folder reactive_widgets
-export 'src/exports/reactive_widgets/reactive_destination.dart';
-export 'src/exports/reactive_widgets/reactive_row_col.dart';
-export 'src/exports/reactive_widgets/reactive_home.dart';