+++ lib/src/exports/extensions/flutter-widgets/text_ext.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/widgets.dart';
-
-import '../../../../xlerate.dart';
+part of '../flutter_widget_ext.dart';
 
 extension TextExt on Text {
   TextCreator get txt => TextCreator(data: data!, style: style);