+++ lib/src/exports/extensions/flutter_widget_ext.dart
@@ -1,0 +1,17 @@
+library flutter_widget_ext;
+
+import 'package:flutter/material.dart';
+import 'package:xlerate/xlerate.dart';
+
+part 'flutter-widgets/align_cntr_ext.dart';
+part 'flutter-widgets/aspectratio_ext.dart';
+part 'flutter-widgets/expanded_and_flexible_ext.dart';
+part 'flutter-widgets/fitted_box_ext.dart';
+part 'flutter-widgets/fractionally_box.dart';
+part 'flutter-widgets/limitedbox_ext.dart';
+part 'flutter-widgets/padding_ext.dart';
+part 'flutter-widgets/positioned_ext.dart';
+part 'flutter-widgets/sizedbox_ext.dart';
+part 'flutter-widgets/stba_ext.dart';
+part 'flutter-widgets/text_ext.dart';
+part 'flutter-widgets/ontap_ext.dart';