+++ lib/xlerate.dart
@@ -1,4 +1,4 @@
-///'simple_sizer' is the core dependency of this package and is usually imported for `extensions` `text-creators` and etc.
+///'xlerate' is the core dependency of this package and is usually imported for `extensions` `text-creators` and etc.
 ///mainly used in android and ios apps [Mobile].
 library xlerate;
 
@@ -13,26 +13,21 @@ export 'src/exports/enums/text_align_enum.dart';
 export 'src/exports/enums/align_cntr_enum.dart';
 
 //From Folder flutter-widgets
-export 'src/exports/extensions/flutter-widgets/build_ctx.dart';
-export 'src/exports/extensions/flutter-widgets/align_cntr_ext.dart';
-export 'src/exports/extensions/flutter-widgets/expanded_and_flexible_ext.dart';
-export 'src/exports/extensions/flutter-widgets/fractionally_box.dart';
-export 'src/exports/extensions/flutter-widgets/padding_ext.dart';
-export 'src/exports/extensions/flutter-widgets/aspectratio_ext.dart';
-export 'src/exports/extensions/flutter-widgets/sizedbox_ext.dart';
-export 'src/exports/extensions/flutter-widgets/fitted_box_ext.dart';
-export 'src/exports/extensions/flutter-widgets/limitedbox_ext.dart';
-export 'src/exports/extensions/flutter-widgets/stba_ext.dart';
-export 'src/exports/extensions/flutter-widgets/text_ext.dart';
-export 'src/exports/extensions/flutter-widgets/positioned_ext.dart';
+export 'src/exports/extensions/flutter_widget_ext.dart';
+export 'src/exports/extensions/build_ctx.dart';
 
 //From Folder data-types
-export 'src/exports/extensions/primitive-types/bool_ext.dart';
-export 'src/exports/extensions/primitive-types/string_ext.dart';
-export 'src/exports/extensions/primitive-types/int_ext.dart';
+export 'src/exports/extensions/data-types/bool_ext.dart';
+export 'src/exports/extensions/data-types/string_ext.dart';
+export 'src/exports/extensions/data-types/int_ext.dart';
+export 'src/exports/extensions/data-types/connection_state_ext.dart';
+export 'src/exports/extensions/data-types/date_time_ext.dart';
 
 //From Folder helper_methods
 export 'src/exports/helper_methods/border_rad_helper.dart';
 
 //Size Config webs
 export 'src/exports/responsiveness/size_config.dart';
+
+//From reactive widgets
+export 'src/exports/reactive_widgets/reactive_btn.dart';